+++ NEWS.md
@@ -1,6 +1,10 @@
 News
 =====
 
+# volcano3D 1.2.0
+###### 25/02/2021
+* Add argument for scene camera in volcano3D
+
 # volcano3D 1.1.0
 ###### 04/02/2021
 * allow colour coding to be based on pvalue or adjusted pvalue according to cutoff_criteria